Refactoring class color into its own file
